packages/indexers/substrate/money_flow/money_flow_indexer_polkadot.py
from loguru import logger
from neo4j import Driver

from packages.indexers.substrate.money_flow.money_flow_indexer import BaseMoneyFlowIndexer


class PolkadotMoneyFlowIndexer(BaseMoneyFlowIndexer):
    """
    Polkadot-specific implementation of the MoneyFlowIndexer.
    This is currently a placeholder that will be implemented in a future step.
    """
    
    def __init__(self, graph_database: Driver, network: str):
        """
        Initialize the PolkadotMoneyFlowIndexer.
        
        Args:
            graph_database: Neo4j driver instance
            network: Network identifier (e.g., 'polkadot')
        """
        super().__init__(graph_database, network)
    
    def _process_network_specific_events(self, transaction, timestamp, events_by_type):
        """
        Process Polkadot-specific events.
        This is currently a placeholder that will be implemented in a future step.
        
        Args:
            transaction: Neo4j transaction
            timestamp: Block timestamp
            events_by_type: Dictionary of events grouped by type
        """
        # TODO: Implement Polkadot-specific event processing
        # For example:
        # self._process_nomination_events(transaction, timestamp, events_by_type.get('Staking.Nomination', []))
        pass