requirements.txt
pydantic
python-dotenv
loguru
pytest
clickhouse-connect
substrate-interface
scalecodec
xxhash
neo4j-rust-ext
fastapi
fastapi-mcp
mcp
uvicorn[standard]
redis
pytest
torch
numpy
fastmcp

